Extract code to override AR::CA::AbstractAdapter.inherited for testing
